Use correct json string for conversion tests
All of the doc options where rolled into a unMeta object in
[version 1.12](https://github.com/jgm/pandoc-types/commit/d325b026e80a5d86bdc261337853f33c02d01317)

Similarily, all JSON data started to be stored as tags and values in
[version 1.12](https://github.com/jgm/pandoc-types/commit/83d600b57f3026b5ccfb0ceaaa8245ebcebcfb27)

As a result, the JSON string that is being tested against no longer
reflects Pandoc output. This change updates the JSON string to match
output from Pandoc 1.13.2.