Remove multiple executable options and replace `bin_path` with `pandoc_path`
The extra scripts that came with previous versions of pandoc were
removed long ago (markdown2pdf was removed in 1.9, html2markdown was
removed in 1.5), so this functionality has been obsolete.

Since it could still be desirable to set a custom path for the `pandoc`
executable, this can be specified using the new `#pandoc_path=` method.
This is similar to the old `#bin_path` method, but it takes a full path
to the `pandoc` executable.

Remove the `#bin_path=` method in favor of the new `#pandoc_path=`
configuration method.