`require 'timeout'` in pandoc-ruby.rb.
Without this `require` statement, tests pass, but the following error is
returned when attempting to perform the same operations in irb/pry:

    uninitialized constant PandocRuby::Timeout

Since there are no other `require 'timeout'` statements, it's unclear
where `timeout` is being required in the tests, so presumably it's
included via one of the testing dependencies.

Currenty it's unclear to me how to write a test for this change due to
this difference likely involving the test dependencies.

refs gh-16