Refactor timeout error message construction.
Use guard to assign a `maybe_ex` variable and include in `error` with
string interpolation in order to keep the lines shorter and hopefully
make the steps slightly clearer.

This commit also appends the contents of `ex` (via the `maybe_ex`
variable) to `error` rather than appending `error` to itself. This
change appears to fulfill the original intent of the `error`
string's construction.

refs gh-16